pyproject.toml
[project]
name = "voicetosrt"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10,<3.11"
dependencies = [
"openai-whisper",
"gradio==5.47.2",
"yt-dlp",
"librosa",
"soundfile",
"resampy",
]

[project.optional-dependencies]
cuda = [
    "torch"
]
xpu = [
    "torch==2.7.0",
    "torchvision==0.22.0",
    "torchaudio==2.7.0",
    "intel-extension-for-pytorch @ https://intel-optimized-pytorch.s3.cn-north-1.amazonaws.com.cn/ipex_stable/xpu/./intel_extension_for_pytorch-2.7.10%2Bxpu-cp310-cp310-win_amd64.whl ",
]

[tool.uv.pip]
index-url = "http://mirrors.tuna.tsinghua.edu.cn/pypi/web/simple"
extra-index-url = [
  "https://download.pytorch.org/whl/xpu",
  # Intel 源是目录索引，uv 不一定按 PEP503 解析；IPEX 改用直链 wheel（见上方 xpu extra）。
  "https://pytorch-extension.intel.com/release-whl/stable/xpu/cn/",
]